added interface and validation with hibernate annotations

+++ src/main/java/com/spring/controller/HelloWorldService.java
@@ -5,17 +5,20 @@ import main.java.com.spring.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 
 @RestController
 @RequestMapping(value="/rest")
-public class HelloWorldService {
+public class HelloWorldService implements IRestController {
 
     @Autowired
     private UserService userService;
@@ -40,7 +43,9 @@ public class HelloWorldService {
 
     //rest/removeCustomer
     @RequestMapping(value="/removeCustomer", method = RequestMethod.POST)
-    public ResponseEntity getCustomerByAge(@RequestBody User user) {
+    public ResponseEntity removeCustomer(@Valid @RequestBody User user, Errors errors) {
+        List<String> errorsArr = processErrors(errors);
+        if (null != errorsArr) return new ResponseEntity(errorsArr, HttpStatus.OK);
         Map<String, Boolean> status = new HashMap<String, Boolean>();
         if(user instanceof User){
             status.put("status", userService.delete(user));
@@ -50,5 +55,15 @@ public class HelloWorldService {
         }
         return new ResponseEntity(status, HttpStatus.OK);
 
+    }
+
+    public List<String> processErrors(Errors errors){
+        if (null != errors && errors.hasErrors()){
+            return errors.getAllErrors()
+                    .stream()
+                    .map(e -> e.getDefaultMessage())
+                    .collect(Collectors.toList());
+        }
+        return null;
     }
 }
